Refactor: move conversions into different files
